requirements.txt
scikit-learn>=1.2.0
numpy>=1.23.0
pandas>=2.0.0
pdfminer.six>=20221105
python-docx>=1.0.0
